test: extract skills e2e fixtures
